reuse Aurora cluster and delete it after tests

+++ wrapper/src/test/java/integration/host/AuroraPostgresContainerTest.java
@@ -95,6 +95,10 @@ public class AuroraPostgresContainerTest {
 
   private static Network network;
   private static final boolean TEST_WITH_EXISTING_DB = EXISTING_DB_CONN_SUFFIX != null;
+  private static final boolean DELETE_EXISTING_DB =
+      !StringUtils.isNullOrEmpty(System.getenv("DELETE_EXISTING_DB"))
+        ? Boolean.parseBoolean(System.getenv("DELETE_EXISTING_DB"))
+        : true;
 
   private static final ContainerHelper containerHelper = new ContainerHelper();
   private static final AuroraTestUtility auroraUtil = new AuroraTestUtility(AURORA_POSTGRES_DB_REGION);
@@ -166,14 +170,16 @@ public class AuroraPostgresContainerTest {
 
   @AfterAll
   static void tearDown() {
-    if (!TEST_WITH_EXISTING_DB) {
-      if (StringUtils.isNullOrEmpty(AURORA_POSTGRES_CLUSTER_IDENTIFIER)) {
-        auroraUtil.deleteCluster();
-      } else {
-        auroraUtil.deleteCluster(AURORA_POSTGRES_CLUSTER_IDENTIFIER);
-      }
+    if (DELETE_EXISTING_DB) {
+      if (!TEST_WITH_EXISTING_DB) {
+        if (StringUtils.isNullOrEmpty(AURORA_POSTGRES_CLUSTER_IDENTIFIER)) {
+          auroraUtil.deleteCluster();
+        } else {
+          auroraUtil.deleteCluster(AURORA_POSTGRES_CLUSTER_IDENTIFIER);
+        }
 
-      auroraUtil.ec2DeauthorizesIP(runnerIP);
+        auroraUtil.ec2DeauthorizesIP(runnerIP);
+      }
     }
 
     for (ToxiproxyContainer proxy : proxyContainers) {
